Update defunct backup URLs
Since keybase.pub was shuttered, you have to use the keybase client to
access the files which is a higher barrier to entry.

It was decided to host the files from the mhct.win website. This commit
will update all keybase URLs to point to the new backups.mhct.win URLs.